Version 0.3.1: Features: A: Server-sent button support added (up to 3 at a time). A: Auction overlay with state info and 6 buttons: [ +1 | +10 | +50 | +100 | + ]. The last button allows you to choose a custom amount to bet or raise. A: Icon support. Trying out Jack's tax and railroad icons. C: Updated the turn info in the center of the board. F: Fixed tapping to close an overlay activating anything underneath. F: Fixed estates with no color gradient appearing with a black name in their overlay. F: Fixed freezing after the game goes out of focus and back and a move animation occurs during that time. F: Fixed a bug where the back button may not work while the game is running. F: Fixed the spelling of "bankruptcy"
Code changes:
A: RotateDrawable added since the android one isn't simple and would require an XML file per bitmap icon and to rotate simple drawables like the house PathShape ShapeDrawable would require me to define it in XML as well.
C: Changed all uses of abstract type java.util.List to java.util.ArrayList.
R: Removed unused code for chat items that was to be used to put an underline in chat when buttons were to be cleared.

git-svn-id: http://monopdroid.googlecode.com/svn/trunk@14 547c2936-0445-ec29-bcaf-865f2df33b66