Version 0.4.0: 8 players, performance, fixes.
Attempted to fix piece animations.
Disabled icon drawing for now, it's really wrong.
Added support for up to 8 players.
Began adding new <configupdate> object support. The protocol seems to
use both though, so keeping old <option> support.
Standardized the location for generating HTML for the overlays for
Player and Estate references. Now all references to a player will use
their piece color and all references to an estate will use it's gradient
color (or white if there is none for either).
The Estate overlay now lists the players on that estate.
Greatly decreased the amount of drawing it does, decreased the number of
animation steps on piece move in an attempt to reduce the amount of
power consumed by the game, especially when nothing is happening.
The Board shallkindly ask your Android device to keep the screen awake.