0.5.0: Add internal support for trades, NYI in UI
Moved attribute lists to their objects.
Configurables are now treated in a way closer to Estates, Players,
Auctions, and Trades.
Total assets is now displayed in the Player overlay (as calculated by
this app).
Attempted to fix the total estate groups field.
Estates overlay now shows the current rent, and will properly display it
for Utilities and Railroads (this is a hack specific to the "city"
gametype and will fail miserably if the group IDs change).
Fixed the roll button not always appearing at the right time.
Added internal support for selling estates back to the bank.
Taps no longer wait for double taps on the board, speeding up perceived
tap-speed in exchange for single taps on estates taking precedence over
double taps to zoom in.
Added "(tap anywhere to close overlay)" helper text for open overlays.
Slightly decreased font size of the BoardView.