Many bug fixes to game join & state
Fixed a crash when joining a game.
Fixed current people in a joined game not being visible.
Fixed many possible crashes & inconsistencies when objects (Player,
Estate, EstateGroup, Auction, or Configurable) updates occur.
Fixed new-style configurables not appearing and a configurable drawing
issue.
Added the "end" game status for when a game ends, so that when it
actually ends it doesn't lead to the red "An error has occured" text
replacing the game board.
Added the game # to the game list.
Added the game # and type to the BoardActivity window title.