Added auto rolling and chat timestamps.
Added the setting to automatically roll the dice.
Added the setting to show timestamps next to all messages.

Fixed the animateMove algorithm so that it does not take a really long
time to calculate and redraw as if the sub-estate animation were still
enabled. Now it uses a timer to act like GtkAtlantic for piece moving.